// src/app/models/VM/vmProductImport.ts
export class VMProductImport {
    ProductName: string;
    Description: string;
    Category: string;
    Brand: string;
    StockQuantity: number = 0;
    AlertQuantity: number = 0;
    PurchasePrice: number = 0;
    SellingPrice: number = 0;
    Unit: string;
    ExpireDate: string;
    Discount: number = 0;
    ProfitMargin: number = 0;
}